+++ CHANGELOG.md
@@ -3,8 +3,9 @@ All notable changes to the "timing" extension will be documented in this file. T
 
 ## Unreleased
 * **DEPRECATED** the `localize` option for advanced `timing.hoverTimestamp.targetFormat`s. Use the new `timezone` option instead.
-* **Added** `timezone` option for advanced `timing.hoverTimestamp.targetFormat`s to enable custom timezone support.
-* **Added** `timing.isoToEpochTimezone` command to enable time conversion including a custom timezone.
+* **Added** `timezone` option for advanced `timing.hoverTimestamp.targetFormat`s to enable custom timezone support. ([#41](https://github.com/HaaLeo/vscode-timing/issues/41))
+* **Added** `timing.epochToIsoTimezone` command to enable ISO time conversion including a custom timezone.
+* **Added** `timing.epochToCustomTimezone` command to enable custom time conversion including a custom timezone.
 
 ## 2020-03-23 - v2.3.1
 * **Fixed** a bug that caused the `timing.nowAsIsoUtc` command to be never shown in the command palette.